Fix 32->16 surrogate handling. Fix range adaptor partial input.
There was a bug in the handling of surrogate values in the
UTF-32->UTF-16 conversion. The code dealt with the <= 0xFFFF
case *before* considering the surrogate case (the high/low surrogate
values are <= 0xFFFF)!

The range adaptor had a bug in the handling of an incomplete code-unit
sequence at the end of the input this could result in a code-unit of
zero being produced. We now consider the end-of-input case after
consuming the input bytes to avoid this problem. Added some additional
assertions.

Added new, more thorough, testing for the UTF-32 to 8/16/32 cases.