Ensure range adaptor cannot reset well_formed_ to true.
It should not be possible for the range adaptor to set the
well_formed_ field to true except during construction. When badly
formed input is detected it is set false, but we should not reset it
to true when processing good input!

Minor renaming, comment typo corrections and an extra assertion.