Add the "byte transcoder" (#70)
In contrast with the existing transcoders which require the source and destination encodings to be known at compile time, the byte transcoder deduces the encoding from the input stream. It determines whether the encoding is one of UTF-8, UTF-16 BE, UTF-16 LE, UTF-32 BE, or UTF-32 LE and converts it to the desired destination encoding.

Improved the UTF-16 transcoder so that an input of HS1,HS2,LS2 is interpreted as REPLACEMENT CHAR, HS2+LS2. i.e. the code unit pair HS2,LS2 is respected. Previously this would have been converted to HS1,LS2.