Fix range iterator copy and move assignment.
We must reseat the valid_ subrange field of the range iterator when an
instance is copied or moved to ensure that the iterators point to the
new object and not the old. Add state copy/move ctor/assign members to
achieve this. The bug was exposed in the Windows release build.

Fix compile failure for the view sentinal class.

Add the RangesUtf16BE test.

Remove the specializations of iter_move and iter_swap which are not necessary.