transcoder<>::partial() is a non-static const member function. (#160)
There were three instances where transcoder<>::partial() was a static
member function because the function simply returns a
constant. However, the transcoder interface has this as a non-static
const member function and I think it's important that all of the
implementations have exactly the same interface. This change may cause
some static analysis warnings, but they can be ignored.